// lib/Q/sys/Corelib.d.ts
export declare namespace string {
    function IsPrintable(keyCode: number, charCode: number): boolean;
}
export declare namespace helper {
    function TryCatch<T>(owner: any, Try: (...args: any[]) => T, Catch?: (e: Error, ...args: any[]) => T, params?: any[]): T;
}
export declare namespace basic {
// ...
    interface IEventHandler extends IDisposable {
        Started: boolean;
        Start(): any;
        Pause(): any;
        Dispose(): any;
        Reset(): any;
    }
    interface Module {
    }
    interface IContext {
        CanAccessToMe(type: string, folder: string, name: string): any;
        GetPath(path: string): string;
        NameOf(type: Function): string;
        GetType(path: string): Function;
    }
    interface IDisposable {
        Dispose(force?: boolean): any;
    }
// ...
    interface IJob {
        Name: string;
        Todo?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        Check?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnError?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
    }
// ...
}
// ...
export declare namespace bind {
    class Job implements basic.IJob {
        Name: string;
        Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        Check: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnError: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnInitialize: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnScopDisposing: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        constructor(Name: string, Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void, Check?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnError?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnInitialize?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnScopDisposing?: (job: JobInstance, e: bind.EventArgs<any, any>) => void);
    }
    class Jobs implements basic.IJob {
        Name: string;
        Todo(job: JobInstance, e: bind.EventArgs<any, any>): void;
        Check(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnError(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing(job: JobInstance, e: bind.EventArgs<any, any>): void;
        constructor(Name: string);
        push(jobName: any): void;
    }
    class JobInstance implements EventListenerObject {
        Scop: bind.Scop;
        job: basic.IJob;
        dom: Node;
        Control: UI.JControl;
        private _events;
        Handle: (ji: JobInstance, e: Event) => void;
        addEventListener(name: string, event: string, delegate: EventListenerOrEventListenerObject | any): void;
        removeEventListener(name: string): void;
        getEvent(name: string): EventListenerOrEventListenerObject;
        constructor(Scop: bind.Scop, job: basic.IJob, dom: Node);
        private propb;
        private ValueChanged(s, e);
        Dispose(): void;
        IsDisposed: any;
        _store: any;
        addValue(name: string, value: any): void;
        getValue(name: string): any;
        Checker: (value: any) => boolean;
        Ischanging: boolean;
        handleEvent(e: Event): void;
    }
    function GetJob(name: string): basic.IJob;
    function Register(job: basic.IJob, override?: boolean): basic.IJob;
}
export declare namespace thread {
    interface IDispatcherCallback {
        callback: (delegate: (...param: any[]) => void, param: any, _this: any) => void;
        params: JobParam;
        _this: any;
        optimizable: boolean;
        isWaiting: boolean;
        id: number;
        children: IDispatcherCallback[];
        ce: number;
    }
    class JobParam {
        id: number;
        params: any[];
        constructor(id: number, params?: any[]);
        Set(...params: any[]): this;
        Set1(params: any[]): this;
        Clone(): JobParam;
    }
    class Dispatcher {
        static OnIdle(owner: any, callback: () => void, once?: boolean): void;
        static InIdle(): boolean;
        static GC(): void;
        static clone(ojob: IDispatcherCallback, params: any[], __this?: any): IDispatcherCallback;
        static cretaeJob(delegate: (...param: any[]) => void, param: any[], _this: any, optimizable: boolean): JobParam;
        static Clear(o: JobParam): void;
        static readonly CurrentJob: IDispatcherCallback;
        private static start();
        static Push(ojob: JobParam, params?: any[], _this?: any): IDispatcherCallback;
        static call(_this: any, fn: Function, ...args: any[]): void;
        static IsRunning(): boolean;
    }
}
// ...
    interface Node<T> {
        Depth: number;
        Value: T;
        param?: any;
        children: Node<T>[];
        Parent: Node<T>;
    }
// ...
export declare namespace Api {
    interface IApiTrigger {
        Name: string;
        Filter: (cmdCallback: IApiCallback, params: any) => boolean;
        CheckAccess: (t: IApiTrigger) => boolean;
        Params?: any;
    }
    interface IApiCallback {
        hash?: string;
        Name: string;
        DoApiCallback: (trigger: IApiTrigger, callback: IApiCallback, params: IApiParam) => void;
        Owner?: any;
        Params?: any;
    }
    function RegisterApiCallback(api: IApiCallback): boolean;
    function RegisterTrigger(api: IApiTrigger): boolean;
    function RiseApi(apiName: string, params: IApiParam): void;
    interface IApiParam {
        data: any;
        callback?(p: IApiParam, args: any): any;
    }
    interface IApi {
        Trigger: IApiTrigger;
        Callback: IApiCallback[];
    }
    interface apiGarbage {
        [name: string]: IApi;
    }
}
// ...
export declare namespace basic {
    class DomEventHandler<T extends Event, P> implements IEventHandler, EventListenerObject {
        dom: Element;
        event: string;
        private owner;
        private handle;
        private param;
        Started: boolean;
        constructor(dom: Element, event: string, owner: any, handle: (eh: DomEventHandler<T, P>, ev: T, param: P) => void, param?: P);
        Start(): void;
        Pause(): void;
        Dispose(): void;
        Reset(): void;
        handleEvent(evt: Event): void;
        static Dispose(dom: EventTarget, event?: string): void;
    }
}
// ...
export declare namespace Notification {
    interface NotificationArgs {
        name: string;
        data: any;
        handler: NotificationHandler<any>;
    }
    interface NotificationHandler<Owner> {
        Id?: any;
        callback: (this: Owner, e: NotificationArgs, ...args) => void;
        owner?: Owner;
        params?: any;
        context?: IContext;
    }
    function on(name: string, handler: NotificationHandler<any>): void;
    function fire(name: string, params: any[]): void;
    function off(name: string, hndl_id: NotificationHandler<any> | any): boolean;
}